@@ -7,7 +7,7 @@
         "esbuild-svelte": "^0.9.0",
         "lodash": "^4.17.21",
         "stylus": "^0.55.0",
-        "svelte": "^4.2.19",
+        "svelte": "^5.1.13",
         "svelte-preprocess": "^6.0.3",
         "typescript": "^5.6.3"
     },
