@@ -1,2 +1,2 @@
 export {getRender} from "./render"
-export {getHooks} from "./hooks"
+export {getHooks} from "./hooks.svelte"
